Remove unnecessary new method and commented out code.
My if statement returns identical results on either side of the condition, but I'm going to throw an error message in there. I can't figure out how to render the page the way the instructions suggest, so I'll just leave it this way for now. It works and the UI is okay. (Form does not repopulate on failed submission.